# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Lesson service",
        "title": "Lesson_sel",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/",
    "paths": {
...
        "/task": {
            "get": {
                "description": "get all tasks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Minutes from",
                        "name": "price_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Minutes to",
                        "name": "price_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/add/{id}": {
            "post": {
                "description": "Add task to lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddTasktoLessonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/img/{id}": {
            "post": {
                "description": "load image to task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UploadImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "description": "get task by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetTaskResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "update task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.TaskItem": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "minutes": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.AddTasktoLessonResponse": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/ds.TaskItem"
                }
            }
        },
...
        "models.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "lesson": {
                    "$ref": "#/definitions/ds.TaskItem"
                }
            }
        },
...
        "models.GetAllTaskResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.TaskItem"
                    }
                },
                "req_id": {
                    "type": "integer"
                },
                "req_less_count": {
                    "type": "integer"
                }
            }
        },
...
        "models.GetTaskResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "$ref": "#/definitions/ds.TaskItem"
                }
            }
        },
...
        "models.UploadImageResponse": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                }
            }
        }
    }
}